DAS-2267: Separate xarray dependencies in the shared_utils directory
