+++ test/nsidc-icesat2/version.txt
@@ -1,1 +1,1 @@
-0.0.4
+0.0.5
